Seperate utility to common
