+++ scripts/gen-tikv-details-dashboard
@@ -16,8 +16,8 @@ docker run \
     -e PYTHONDONTWRITEBYTECODE=1 \
     tikv-dashboard-gen \
     bash -c "\
-        isort /metrics/grafana/tikv_details.dashboard.py && \
-        black /metrics/grafana/tikv_details.dashboard.py && \
+        isort --profile black /metrics/grafana/*.py && \
+        black /metrics/grafana/*.py && \
         generate-dashboard \
             -o /metrics/grafana/tikv_details1.json \
             /metrics/grafana/tikv_details.dashboard.py
